changed omnisearch cutoffs, moved omnisearch client side code to mass/search.ts
